backend/requirements.txt
pymongo
python-dotenv
flask
flask-cors
flask-bcrypt
flask-jwt-extended
flask-sqlalchemy
